requirements.txt
# Core data analysis
pandas>=2.0.0
numpy>=1.24.0

# Financial data
yfinance>=0.2.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0
pykalman>=0.9.7
hmmlearn>=0.3.0
scipy>=1.11.0

# Configuration
PyYAML>=6.0

# Logging
loguru>=0.7.0

# CLI
click>=8.0.0

# Type hints
typing-extensions>=4.0.0

# Development tools (optional for production)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0